requirements.txt
# Core Dependencies - Based on actual project usage
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# AI/ML Libraries 
openai==1.3.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Document Processing
PyPDF2==3.0.1

# Data Processing
numpy==1.25.2

# Web Interface
streamlit==1.28.1
requests==2.31.0

# Utilities
python-multipart==0.0.6

# Development & Testing (Optional)
# Uncomment if you want to use code formatting:
# black==23.11.0